first initial cleanup and refactoring changes
