+++ backend/grocery-mate-backend/grocery-mate-backend.csproj
@@ -25,6 +25,9 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Folder Include="BusinessLogic\Validation\Shopping" />
+    <Folder Include="BusinessLogic\Validation\UserSettings" />
+    <Folder Include="Controllers\Handler" />
     <Folder Include="Data\DataModels" />
     <Folder Include="Migrations" />
   </ItemGroup>
